% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{child_node}
\alias{child_node}
\title{Wrapper to extract sub-nodes within XMLNode objects}
\usage{
child_node(x)
}
\arguments{
\item{x}{An object of class \code{XMLInternalElementNode} obtained from \code{root_node}.}
}
\description{
Wrapper for \code{XML::xmlChildren}. For internal use only.
}
\seealso{
\code{\link{html_parse}}, \code{\link{root_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{extract_summary_tables}}, \code{\link{tidy}}, \code{\link{extract_coverage_stats}}, \code{qualimap}
}

