% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{extract_coverage_stats}
\alias{extract_coverage_stats}
\title{Load Qualimap coverage stats}
\usage{
extract_coverage_stats(coverage_file)
}
\arguments{
\item{coverage_file}{A complete path to the text file containing coverage stats.}
}
\description{
Loads the coverage stats. For internal use only.
}
\seealso{
\code{\link{html_parse}}, \code{\link{root_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{extract_summary_tables}}, \code{\link{tidy}}, \code{qualimap}
}

