% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{extract_summary_tables}
\alias{extract_summary_tables}
\title{Extract all summary tables from Qualimap}
\usage{
extract_summary_tables(doc)
}
\arguments{
\item{doc}{An object of class \code{HTMLInternalDocument} obtained from \code{html_parse}.}
}
\description{
The function extracts just the summary tables from a Qualimap report file on to
a list of \code{data.table}s and tidies them up. For internal use only.
}
\seealso{
\code{\link{root_node}}, \code{\link{child_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{tidy}}, \code{\link{extract_coverage_stats}}, \code{qualimap}
}

