% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggqualimap.R
\docType{package}
\name{ggqualimap}
\alias{ggqualimap}
\alias{ggqualimap-package}
\title{ggqualimap: Plots of Statistics Collected by Qualimap from RNASeq Data}
\description{
ggqualimap: Plots of Statistics Collected by Qualimap from RNASeq Data
}

