% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{html_table}
\alias{html_table}
\title{Extract all tables into list of data frames}
\usage{
html_table(x)
}
\arguments{
\item{x}{An object of class \code{HTMLInternalDocument} obtained from \code{html_parse}.}
}
\description{
Wrapper for \code{XML::readHTMLTable}. For internal use only.
}
\seealso{
\code{\link{html_parse}}, \code{\link{root_node}}, \code{\link{child_node}}, \code{\link{qualimap_tables}}, \code{\link{extract_summary_tables}}, \code{\link{tidy}}, \code{\link{extract_coverage_stats}}, \code{qualimap}
}

