% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-plots.R
\name{plot_genomic_origin}
\alias{plot_genomic_origin}
\title{Plot Qualimap genomic origin}
\usage{
plot_genomic_origin(x, type = c("ggplot2", "rbokeh"))
}
\arguments{
\item{x}{The list generated by running \code{qualimap()} function that has gathered all tables from all samples together.}

\item{type}{Package to use to plot the summaries. Only \code{ggplot2} is implemented for now.}
}
\description{
Takes in the list object generated from \code{qualimap()} function and generates
genomic origin plots.
}
\seealso{
\code{\link{plot_read_alignment}}, \code{\link{plot_bias_profile}}, \code{\link{plot_coverage_profile}}, \code{\link{plot_junction_analysis}}
}

