% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualimap-plots.R
\name{plot_junction_analysis}
\alias{plot_junction_analysis}
\title{Plot Qualimap junction analysis}
\usage{
plot_junction_analysis(..., interactive = TRUE, geom = "bar")
}
\arguments{
\item{interactive}{logical, default is \code{TRUE}, which returns an 
\emph{interactive} \code{plotly} plot. If \code{FALSE}, it returns a 
static \code{ggplot2} plot.}

\item{geom}{Only possible value is \code{"bar"}.}

\item{\dots}{The set of \code{qualimap} objects to plot, usually of the 
form \code{sample_name_1 = obj1}, \code{sample_name_2 = obj2}, etc. See 
\code{examples}. The names will be used as title for facets.}
}
\value{
The plot object
}
\description{
This function plots the junction stats for samples. If 
info regarding the \code{group} each \code{sample} belongs to is also 
available, then the generated plot will take that into account to colour / 
facet accordingly.
}
\examples{
path = system.file("tests/qualimap-sample", package="ggqualimap")
obj = qualimap(sample_info = file.path(path, "annotation.txt"))

# interactive
plot_junction_analysis(sample = obj)

# non-interactive
plot_junction_analysis(sample = obj, interactive=FALSE)
}
\seealso{
\code{\link{qualimap}} \code{\link{plot_read_alignment}}
\code{\link{plot_bias_profile}} \code{\link{plot_coverage_profile}}
\code{\link{plot_genomic_origin}}
}

