% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-plots.R
\name{plot_read_alignment}
\alias{plot_read_alignment}
\title{Plot Qualimap read alignment}
\usage{
plot_read_alignment(x, type = c("ggplot2", "rbokeh"), group = NULL)
}
\arguments{
\item{x}{The list generated by running \code{qualimap()} function that has gathered all tables from all samples together.}

\item{type}{Package to use to plot the summaries. Only \code{ggplot2} is implemented for now.}

\item{group}{Default is \code{NULL}. If the plots need to be grouped by samples, then a two
column \code{data.table} has to be provided with names \code{sample_name} and \code{grp}. The
column \code{grp} should have identical values for all samples belonging to the same group.}
}
\description{
Takes in the list object generated from \code{qualimap()} function and generates
read alignment plots.
}
\seealso{
\code{\link{plot_genomic_origin}}, \code{\link{plot_bias_profile}}, \code{\link{plot_coverage_profile}}, \code{\link{plot_junction_analysis}}
}

