% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{qualimap_tables}
\alias{qualimap_tables}
\title{Extract all summary tables for a particular sample}
\usage{
qualimap_tables(file, sample, type, group)
}
\arguments{
\item{file}{The file corresponding to \emph{that \code{sample}}.}

\item{sample}{Sample name.}

\item{group}{Group / condition this sample belongs to.}

\item{pair}{Read pair.}
}
\description{
\bold{NOTE:} For internal use only.

This is the workhorse that powers \code{qualimap}. Given a \code{sample}, 
\code{type} and \code{path}, \code{extract_summary_tables} extracts all 
summary statistics and returns them as a list of \code{data.table}s.
}

