% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{root_node}
\alias{root_node}
\title{Wrapper to extract top-level XML node}
\usage{
root_node(doc)
}
\arguments{
\item{doc}{An object of class \code{HTMLInternalDocument} obtained from \code{html_parse}.}
}
\description{
Wrapper for \code{XML::xmlRoot}. For internal use only.
}
\seealso{
\code{\link{html_parse}}, \code{\link{child_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{extract_summary_tables}}, \code{\link{tidy}}, \code{\link{extract_coverage_stats}}, \code{qualimap}
}

