% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-ggqualimap.R
\name{test_ggqualimap}
\alias{test_ggqualimap}
\title{Run a set of tests}
\usage{
test_ggqualimap(verbose = FALSE, pkg = "pkg", silent = FALSE)
}
\arguments{
\item{verbose}{If \code{TRUE} sets \code{ggqualimap_verbose} to \code{TRUE} for 
the duration of the tests.}

\item{pkg}{Root directory name under which all package content 
(ex: \code{DESCRIPTION}, \code{src/}, \code{R/}, \code{inst/} etc..) 
resides.}

\item{silent}{Logical, default \code{FALSE}. When \code{TRUE} it will not 
raise error on in case of test fails.}
}
\value{
When silent equals to \code{TRUE} it will return \code{TRUE} if 
all tests were successful. \code{FALSE} otherwise. If silent equals to 
\code{FALSE} it will return \code{TRUE} if all tests were successful. Error 
otherwise.
}
\description{
Runs a set of tests to check if \code{ggqualimap} is working 
correctly.
}
\details{
Runs a series of tests. These can be used to see features and 
examples of usage, too. Running \code{test_ggqualimap} will tell you the full 
location of the test file(s) to open.
}
\examples{
\dontrun{
ggqualimap:::test_ggqualimap()
}
}
\seealso{
\code{\link{qualimap}} \code{\link{plot_read_alignment}} 
\code{\link{plot_bias_profile}} \code{\link{plot_coverage_profile}} 
\code{\link{plot_junction_analysis}} \code{\link{plot_genomic_origin}}
}

