% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{tidy}
\alias{tidy}
\title{Cleanup all summary tables}
\usage{
tidy(x)
}
\arguments{
\item{html_file}{Complete path to html file from Qualimap.}
}
\description{
For internal use only.
}
\seealso{
\code{\link{root_node}}, \code{\link{child_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{extract_summary_tables}}, \code{\link{extract_coverage_stats}}, \code{qualimap}
}

