% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-plots.R
\name{plot_junction_analysis}
\alias{plot_junction_analysis}
\title{Plot Qualimap junction analysis}
\usage{
plot_junction_analysis(x, type = c("ggplot2", "rbokeh"))
}
\arguments{
\item{x}{The list generated by running \code{qualimap()} function that has gathered all tables from all samples together.}

\item{type}{Package to use to plot the summaries. Only \code{ggplot2} is implemented for now.}
}
\description{
Takes in the list object generated from \code{qualimap()} function and generates
junction analysis plots.
}
\seealso{
\code{\link{plot_read_alignment}}, \code{\link{plot_genomic_origin}}, \code{\link{plot_coverage_profile}}, \code{\link{plot_bias_profile}}
}

