% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{qualimap}
\alias{qualimap}
\title{Gather all Qualimap summary tables from all samples}
\usage{
qualimap(dir, groups = NULL)
}
\arguments{
\item{dir}{Complete path to Qualimap reports for all samples from an experiment}

\item{groups}{Default is \code{NULL}. If the plots need to be grouped by samples, then a two
column \code{data.table} has to be provided with names \code{sample} and \code{group}. The
column \code{group} should have identical values for all samples belonging to the same group.}
}
\description{
This function extracts the \code{Input}, \code{Reads alignment},
\code{Reads genomics origin}, \code{Transcript coverage profile}, \code{Junction
Analysis} and \code{coverage} tables from all samples.
}
\seealso{
\code{\link{root_node}}, \code{\link{child_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{tidy}}, \code{\link{extract_coverage_stats}}, , \code{\link{extract_summary_tables}}
}

