% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualimap-tables.R
\name{qualimap_tables}
\alias{qualimap_tables}
\title{Extract all summary tables from Qualimap report}
\usage{
qualimap_tables(sample_name = "", html_file, groups = NULL)
}
\arguments{
\item{sample_name}{Sample name to be appeneded as a separate column to the summary/coverage tables.}

\item{html_file}{Complete path to html file from Qualimap.}

\item{groups}{Default is \code{NULL}. If the plots need to be grouped by samples, then a two
column \code{data.table} has to be provided with names \code{sample} and \code{group}. The
column \code{group} should have identical values for all samples belonging to the same group.}
}
\description{
This function extracts the \code{Input}, \code{Reads alignment},
\code{Reads genomics origin}, \code{Transcript coverage profile}, \code{Junction
Analysis} tables.
}
\examples{
\dontrun{
exp_tables = qualimap_tables("qualimapReport.html", "coverage_profile_along_genes_(total).txt")
}
}
\seealso{
\code{\link{root_node}}, \code{\link{child_node}}, \code{\link{html_table}}, \code{\link{qualimap_tables}}, \code{\link{tidy}}, \code{\link{extract_coverage_stats}}, , \code{\link{extract_summary_tables}}, \code{qualimap}
}

